+++ lib/datadog/appsec/waf/result.rb
@@ -6,20 +6,29 @@ module Datadog
       # Ruby representation of the ddwaf_result of a libddwaf run.
       # See https://github.com/DataDog/libddwaf/blob/8dbee187ff74a0aa25e1bcbdde51677f77930e1b/include/ddwaf.h#L277-L290
       class Result
-        attr_reader :status, :events, :actions, :derivatives, :total_runtime, :timeout
+        attr_reader :status, :events, :actions, :derivatives, :total_runtime
 
-        def initialize(status, events, actions, derivatives, total_runtime, timeout)
+        def initialize(status, events, actions, derivatives, total_runtime, timeout, keep)
           @status = status
           @events = events
           @actions = actions
           @derivatives = derivatives
           @total_runtime = total_runtime
           @timeout = timeout
+          @keep = keep
           @input_truncated = false
         end
 
         def mark_input_truncated!
           @input_truncated = true
+        end
+
+        def timeout?
+          @timeout
+        end
+
+        def keep?
+          @keep
         end
 
         def input_truncated?
@@ -34,6 +43,7 @@ module Datadog
             derivatives: @derivatives,
             total_runtime: @total_runtime,
             timeout: @timeout,
+            keep: @keep,
             input_truncated: @input_truncated
           }
         end