+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Fixed
 
-- Fix conversion of `nil` and Integer Hash keys
+- Fix conversion of non-string Hash keys with potential size truncation
 
 # 2025-09-02 v1.24.1.2.0